refactor(6552): isolate moratorium git-range logic
